% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNEXTlink.R
\name{Asy3D.link}
\alias{Asy3D.link}
\title{Asymptotic diversity q profile}
\usage{
Asy3D.link(
  data,
  class = "TD",
  q = seq(0, 2, 0.2),
  datatype = "abundance",
  nboot = 30,
  conf = 0.95,
  row.tree = NULL,
  col.tree = NULL
)
}
\arguments{
\item{outcome}{the outcome of the functions \code{AsyD} .\cr}
}
\value{
a table of Asymptoti network diversity q profile
}
\description{
\code{Asy3D.link} The estimated and empirical diversity of order q
}
\examples{
\dontrun{
## Ex.1
data(Norfolk)
out1 <- Asy3D.link(Norfolk, class = 'TD', datatype = "abundance", nboot = 10)
ggAsy3D.link(out1)
## Ex.2
data(puerto.rico)
out2 <- Asy3D.link(puerto.rico$data, class = 'PD', datatype = "abundance", nboot = 10, row.tree = puerto.rico$row.tree, col.tree = puerto.rico$col.tree)
ggAsy3D.link(out2)
}

}
