% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNEXTlink.R
\name{AsyND}
\alias{AsyND}
\title{Asymptotic diversity q profile}
\usage{
AsyND(
  data,
  q = seq(0, 2, 0.2),
  datatype = "abundance",
  nboot = 50,
  conf = 0.95
)
}
\arguments{
\item{outcome}{the outcome of the functions \code{AsyD} .\cr}
}
\value{
a table of Asymptoti network diversity q profile
}
\description{
\code{AsyND} The estimated and empirical diversity of order q
}
\examples{
\dontrun{
data(Norfolk)
out1 <- AsyND(Norfolk,datatype = "abundance")
ggAsyPND(out1)
}

}
