% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNEXTlink.R
\name{AsyPND}
\alias{AsyPND}
\title{Asymptotic Phylogenetic diversity q profile}
\usage{
AsyPND(
  data = puerto.rico$data,
  q = seq(0, 2, 0.2),
  datatype = "abundance",
  row.tree = NULL,
  col.tree = NULL,
  nboot = 50,
  conf = 0.95
)
}
\arguments{
\item{outcome}{the outcome of the functions \code{AsyD} .\cr}
}
\value{
a table of Asymptotic network diversity q profile
}
\description{
\code{AsyPND} The estimated and empirical diversity of order q
}
\examples{
\dontrun{
## Type (1) example for abundance-based data
## Ex.1
data(puerto.rico$data)
out1 <- AsyND(puerto.rico$data, datatype = "abundance", row.tree = row.tree,col.tree = col.tree)
ggAsyPND(out1)
}
}
