\name{beetles}
\alias{beetles}
\docType{data}
\title{
Plants-beetle interaction networks
}
\description{
This dataset contains two quantitative plant-beetle interaction networks from two forest subplots (Open and Closed) in the Steigerwald Forest, northern Bavaria, Germany. The data were originally published by Vogel et al. (2020). For each subplot, the interaction frequencies (or strengths) between beetle species and plant species are recorded in a bipartite matrix.\cr
}
\usage{data("beetles")}
\format{
  \code{beetles} is a list with two networks data.frames/matrices. The two list represent
  different habitats/subplots. \cr
  A list of 2 matrices: \cr
  $ Open is a matrix with 238 beetles(rows) and 6 plant (columns). \cr
  $ Closed is a matrix with 238 beetles(rows) and 6 plant (columns). \cr
}
\references{
Vogel, S., et al. (2020). "Optimizing enrichment of deadwood for biodiversity by varying sun   exposure and tree species: an experimental approach." Journal of Applied Ecology 57(10): 2075-2085.
}

\keyword{datasets}
