\name{beetles_plotA}
\alias{beetles_plotA}
\docType{data}
\title{
Plants-beetle interaction network: plot A
}
\description{
This dataset contains a quantitative plant-beetle interaction network sampled from plot A, one of six experimental forest plots established in the Steigerwald Forest, northern Bavaria, Germany (Vogel, 2020).
}
\usage{data("beetles_plotA")}
\format{
  \code{beetles} is a list of two interaction matrices. The two list represent
  different habitats/subplots. \cr
  A list of 2 matrices: \cr
  $ Open is a matrix with 83 beetles(rows) and 6 plant (columns). \cr
  $ Closed is a matrix with 83 beetles(rows) and 6 plant (columns). \cr
}
\references{
Vogel, S., et al. (2020). "Optimizing enrichment of deadwood for biodiversity by varying sun   exposure and tree species: an experimental approach." Journal of Applied Ecology 57(10): 2075-2085.
}

\keyword{datasets}
