\name{beetles_row_tree}
\alias{beetles_row_tree}
\docType{data}
\title{
Phylogenetic tree for beetles data and beetles_plotA data.
}
\description{
This dataset includes a phylogenetic tree spanned by 1449 species listed in the \code{beetles} dataset and \code{beetles_plotA} dataset.
}
\usage{data("beetles_row_tree")}
\format{
 \code{beetles_row_tree} is a list (phylo tree) with the following phylogenetic information: \cr
 A List of 6: \cr
 $ edge       : int [1:2205, 1:2] 1450 1450 1451 1452 1452 1453 1453 1451 1454 1455 ... \cr
 $ edge.length: num [1:2205] 285 8.56 126.42 150.02 126.42 ... \cr
 $ Nnode      : int 757 \cr
 $ node.label : chr [1:757] "match0" "match1" "match2" "match3" ... \cr
 $ tip.label  : chr [1:1449] "Neuroptera" "Tachyta_nana" "Rhysodes_sulcatus"  "Omoglymmius_germari" \cr
}
\references{
Vogel, S., et al. (2020). "Optimizing enrichment of deadwood for biodiversity by varying sun   exposure and tree species: an experimental approach." Journal of Applied Ecology 57(10): 2075-2085.
}
\keyword{datasets}
