% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNEXTlink.R
\name{DataInfo.link}
\alias{DataInfo.link}
\title{Exhibit basic data information}
\usage{
DataInfo.link(
  data,
  class,
  datatype = "abundance",
  row.tree = NULL,
  col.tree = NULL
)
}
\arguments{
\item{data}{a vector/matrix/list of species abundances or incidence frequencies.\cr If \code{datatype = "incidence"},
then the first entry of the input data must be total number of sampling units, followed by species incidence frequencies.}

\item{class}{a choice of three-level diversity: 'TD' = 'Taxonomic', 'PD' = 'Phylogenetic', and 'FD' = 'Functional' under certain threshold.}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}),
sampling-unit-based incidence frequencies data (\code{datatype = "incidence_freq"}) or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}).}
}
\value{
a data.frame of basic data information including sample size, observed species richness, sample coverage estimate, and the first ten abundance/incidence frequency counts.
}
\description{
\code{DataInfo.link}: exhibits basic data information
}
\examples{
\dontrun{
data(puerto.rico)
DataInfo.link(puerto.rico$data, class = 'TD', datatype="abundance")
DataInfo.link(puerto.rico$data, class = 'PD', datatype="abundance",
row.tree = puerto.rico$row.tree, col.tree = puerto.rico$col.tree)
}
}
