% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNEXTlink.R
\name{estimatePND}
\alias{estimatePND}
\title{Compute species diversity with a particular of sample size/coverage}
\usage{
estimatePND(
  data,
  q = c(0, 1, 2),
  datatype = "abundance",
  row.tree = NULL,
  col.tree = NULL,
  level = NULL,
  nboot = 30,
  conf = 0.95
)
}
\arguments{
\item{outcome}{the outcome of the functions \code{ObsND} .\cr}
}
\value{
a data.frame of species diversity table including the sample size, sample coverage, method (rarefaction or extrapolation), and diversity estimates with q = 0, 1, and 2 for the user-specified sample size or sample coverage.
}
\description{
\code{estimatedPND} computes species diversity (Hill numbers with q = 0, 1 and 2) with a particular user-specified level of sample size or sample coverage.
}
\examples{
\dontrun{
data(puerto.rico$data)
out <- estimatedPND(puerto.rico$data, q = c(0,1,2), datatype = "abundance", row.tree = rowtree, col.tree = coltree)
out
}

}
