% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNEXTlink.R
\name{ggAsyND}
\alias{ggAsyND}
\title{ggplot for Asymptotic Network diversity}
\usage{
ggAsyND(outcome)
}
\arguments{
\item{outcome}{the outcome of the functions \code{AsyND} .\cr}
}
\value{
a figure of estimated sample completeness with order q\cr\cr
}
\description{
\code{ggAsyND} Plots q-profile based on the outcome of \code{AsyND} using the ggplot2 package.\cr
It will only show the confidence interval of 'Estimated'.
}
\examples{
## example for abundance-based data
## Ex.1
data(Norfolk)
out1 <- AsyND(Norfolk, datatype = "abundance")
ggAsyND(out1)
}
