% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_function.R
\name{ggCompleteness.link}
\alias{ggCompleteness.link}
\title{ggplot for Sample Completeness}
\usage{
ggCompleteness.link(outcome)
}
\arguments{
\item{outcome}{a table generated from Completeness.link function}
}
\value{
a figure of estimated sample completeness with order q
}
\description{
\code{ggCompleteness.link} The figure for estimation of Sample Completeness with order q
}
\examples{
data(beetles)
output = Completeness.link(beetles)
ggCompleteness.link(output)
}
