% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNEXTlink.R
\name{ggiNEXT.link}
\alias{ggiNEXT.link}
\title{ggplot2 extension for an iNEXT object}
\usage{
ggiNEXT.link(
  outcome,
  class = "TD",
  type = 1,
  se = TRUE,
  facet.var = "Assemblage",
  color.var = "Order.q",
  text.size = 18
)
}
\arguments{
\item{type}{three types of plots: sample-size-based rarefaction/extrapolation curve (\code{type = 1});
sample completeness curve (\code{type = 2}); coverage-based rarefaction/extrapolation curve (\code{type = 3}).}

\item{se}{a logical variable to display confidence interval around the estimated sampling curve.}

\item{facet.var}{create a separate plot for each value of a specified variable:
no separation \cr (\code{facet.var="None"});
a separate plot for each diversity order (\code{facet.var="Order.q"});
a separate plot for each assemblage (\code{facet.var="Assemblage"});
a separate plot for each combination of order x assemblage (\code{facet.var="Both"}).}

\item{color.var}{create curves in different colors for values of a specified variable:
all curves are in the same color (\code{color.var="None"});
use different colors for diversity orders (\code{color.var="Order.q"});
use different colors for sites (\code{color.var="Assemblage"});
use different colors for combinations of order x assemblage (\code{color.var="Both"}).}

\item{x}{an \code{iNEXT} object computed by \code{\link{iNEXT}}.}

\item{grey}{a logical variable to display grey and white ggplot2 theme.}

\item{...}{other arguments passed on to methods. Not currently used.}
}
\value{
a ggplot2 object
}
\description{
\code{ggiNEXT.link}: the \code{\link[ggplot2]{ggplot}} extension for \code{\link{iNEXT}} Object to plot sample-size- and coverage-based rarefaction/extrapolation curves along with a bridging sample completeness curve
}
\examples{
\dontrun{
data(Norfolk)
out1 <- iNEXT.link(Norfolk, class = 'TD',datatype = "abundance", nboot = 0)
ggiNEXT.link(outcome = out1, type = 1)
ggiNEXT.link(outcome = out1, type = 2)
ggiNEXT.link(outcome = out1, type = 3)

#' data(puerto.rico)
out2 <- iNEXT.link(puerto.rico$data, class = 'PD', datatype="abundance",
row.tree = puerto.rico$row.tree, col.tree = puerto.rico$col.tree)
ggiNEXT.link(outcome = out2, type = 1)
ggiNEXT.link(outcome = out2, type = 2)
ggiNEXT.link(outcome = out2, type = 3)
}
}
