% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNEXTlink.R
\name{ggiNEXT_beta_link}
\alias{ggiNEXT_beta_link}
\title{ggplot2 extension for an iNEXT object}
\usage{
ggiNEXT_beta_link(
  output,
  type = c("B", "D"),
  measurement = c("T", "P", "F_tau", "F_AUC"),
  scale = "free",
  main = NULL,
  transp = 0.4
)
}
\arguments{
\item{type}{three types of plots: sample-size-based rarefaction/extrapolation curve (\code{type = 1});
sample completeness curve (\code{type = 2}); coverage-based rarefaction/extrapolation curve (\code{type = 3}).}

\item{x}{an \code{iNEXT} object computed by \code{\link{iNEXT}}.}

\item{se}{a logical variable to display confidence interval around the estimated sampling curve.}

\item{facet.var}{create a separate plot for each value of a specified variable:
no separation \cr (\code{facet.var="None"});
a separate plot for each diversity order (\code{facet.var="Order.q"});
a separate plot for each assemblage (\code{facet.var="Assemblage"});
a separate plot for each combination of order x assemblage (\code{facet.var="Both"}).}

\item{color.var}{create curves in different colors for values of a specified variable:
all curves are in the same color (\code{color.var="None"});
use different colors for diversity orders (\code{color.var="Order.q"});
use different colors for sites (\code{color.var="Assemblage"});
use different colors for combinations of order x assemblage (\code{color.var="Both"}).}

\item{grey}{a logical variable to display grey and white ggplot2 theme.}

\item{...}{other arguments passed on to methods. Not currently used.}
}
\value{
a ggplot2 object
}
\description{
\code{ggiNEXT_beta_link}: the \code{\link[ggplot2]{ggplot}} extension for
\code{\link{iNEXTlink}} Object to plot sample-size- and coverage-based rarefaction/extrapolation curves along with a bridging sample completeness curve
}
\examples{
\dontrun{
data(Norfolk)
out = iNEXT_beta_link(Norfolk, coverage_expected, data_type=c('abundance', 'incidence_raw'), q = c(0, 1, 2),
level=c('taxonomic', 'phylogenetic', 'functional'), nboot = 20, conf = 0.95, max_alpha_coverage=F,
 by=c('coverage', 'size'),phy_tree=NULL, reftime = NULL)

ggiNEXT_beta_link(out, type = 1,se = TRUE,facet = "None",color = "Assemblage",grey = FALSE, text_size = 10)
}
}
