% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNEXTlink.R
\name{ggiNEXT_PND}
\alias{ggiNEXT_PND}
\title{ggplot2 extension for an iNEXT object}
\usage{
ggiNEXT_PND(
  outcome,
  type = 1,
  se = TRUE,
  facet = "None",
  color = "Assemblage",
  grey = FALSE,
  text_size = 10
)
}
\arguments{
\item{type}{three types of plots: sample-size-based rarefaction/extrapolation curve (\code{type = 1});
sample completeness curve (\code{type = 2}); coverage-based rarefaction/extrapolation curve (\code{type = 3}).}

\item{se}{a logical variable to display confidence interval around the estimated sampling curve.}

\item{grey}{a logical variable to display grey and white ggplot2 theme.}

\item{x}{an \code{iNEXT} object computed by \code{\link{iNEXT}}.}

\item{facet.var}{create a separate plot for each value of a specified variable:
no separation \cr (\code{facet.var="None"});
a separate plot for each diversity order (\code{facet.var="Order.q"});
a separate plot for each assemblage (\code{facet.var="Assemblage"});
a separate plot for each combination of order x assemblage (\code{facet.var="Both"}).}

\item{color.var}{create curves in different colors for values of a specified variable:
all curves are in the same color (\code{color.var="None"});
use different colors for diversity orders (\code{color.var="Order.q"});
use different colors for sites (\code{color.var="Assemblage"});
use different colors for combinations of order x assemblage (\code{color.var="Both"}).}

\item{...}{other arguments passed on to methods. Not currently used.}
}
\value{
a ggplot2 object
}
\description{
\code{ggiNEXT}: the \code{\link[ggplot2]{ggplot}} extension for \code{\link{iNEXTlink}} Object to plot sample-size- and coverage-based rarefaction/extrapolation curves along with a bridging sample completeness curve
}
\examples{
data(spider)
# single-assemblage abundance data
out1 <- iNEXT_ND(spider$Girdled, q = 0, datatype = "abundance")
ggiNEXT_ND(x = out1, type = 1)
ggiNEXT_ND(x = out1, type = 2)
ggiNEXT_ND(x = out1, type = 3)

# single-assemblage incidence data with three orders q
data(ant)
size <- round(seq(10, 500, length.out=20))
y <- iNEXT_ND(ant$h500m, class = 'TD', q = c(0,1,2), datatype = "incidence_freq", size = size)
ggiNEXT_ND(y, se=FALSE, color.var="Order.q")

# multiple-assemblage abundance data with three orders q
data(spider)
z <- iNEXT_ND(spider, class = 'TD', q = c(0,1,2), datatype = "abundance")
ggiNEXT_ND(z, facet.var="Assemblage", color.var="Order.q")
ggiNEXT_ND(z, facet.var="Both", color.var="Both")
}
