% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNEXTlink.R
\name{ggNetEven}
\alias{ggNetEven}
\title{ggplot for Evenness
\code{ggNetEven} The figure for estimation of Evenness with order q\cr}
\usage{
ggNetEven(output)
}
\arguments{
\item{output}{a table generated from Evenness function\cr}
}
\value{
a figure of estimated sample completeness with order q\cr
}
\description{
ggplot for Evenness
\code{ggNetEven} The figure for estimation of Evenness with order q\cr
}
\references{
Chao,A.and Ricotta,C.(2019).Quantifying evenness and linking it to diversity, beta diversity, and similarity.
}
