% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNEXTlink.R
\name{ggObs.link}
\alias{ggObs.link}
\title{ggplot for Empirical Network diversity}
\usage{
ggObs.link(outcome, text.size = 14)
}
\arguments{
\item{outcome}{the outcome of the functions \code{Obs.link} .\cr}

\item{text.size}{control the text size of the output plot.}
}
\value{
a figure of estimated sample completeness with order q\cr\cr
}
\description{
\code{ggObs.link} Plots q-profile based on the outcome of \code{Obs.link} using the ggplot2 package.\cr
}
\examples{
\dontrun{
## Ex.1
data(Norfolk)
out1 <- Obs.link(Norfolk, diversity = 'TD', datatype = "abundance", nboot = 10)
ggObs.link(out1)
## Ex.2
data(puerto.rico)
out2 <- Obs.link(puerto.rico$data, diversity = 'PD', datatype = "abundance",
nboot = 10, row.tree = puerto.rico$row.tree, col.tree = puerto.rico$col.tree)
ggObs.link(out2)
}
}
