% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{ggObsAsy.link}
\alias{ggObsAsy.link}
\title{ggplot for Asymptotic Network diversity}
\usage{
ggObsAsy.link(output)
}
\arguments{
\item{output}{the output of the functions \code{ObsAsy.link} .\cr}
}
\value{
a figure of asymptotic or empirical (observed) diversity in q-profile.\cr\cr
}
\description{
\code{ggObsAsy.link} Plots q-profile based on the output of \code{ObsAsy.link} using the ggplot2 package.\cr
}
\examples{
\donttest{
# Plot q-profile of taxonomic network diversity for interaction data
# with order q between 0 and 2 (in increments of 0.2 by default).
data(beetles_plotA)
output_ObsAsy_qiTD = ObsAsy.link(data = beetles_plotA, diversity = 'TD', q = seq(0, 2, 0.2))
ggObsAsy.link(output_ObsAsy_qiTD)


# Plot q-profile of phylogenetic network diversity for interaction data
# with order q between 0 and 2 (in increments of 0.2 by default).
data(beetles_plotA)
data(beetles_row_tree)
output_ObsAsy_qiPD = ObsAsy.link(data = beetles_plotA, diversity = 'PD', q = seq(0, 2, 0.2), 
                                 row.tree = beetles_row_tree, nboot = 10)
ggObsAsy.link(output_ObsAsy_qiPD)


# Plot q-profile of functional network diversity for interaction data
# with order q between 0 and 2 (in increments of 0.2 by default)
# under tau values from 0 to 1
data(beetles_plotA)
data(beetles_row_distM)
output_ObsAsy_qiFD = ObsAsy.link(data = beetles_plotA, diversity = 'FD', q = seq(0, 2, 0.25), 
                                 row.distM = beetles_row_distM, FDtype = "AUC", nboot = 10)
ggObsAsy.link(output_ObsAsy_qiFD)
}

}
