% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNEXTlink.R
\name{ggSC.link}
\alias{ggSC.link}
\title{ggplot for Sample Completeness}
\usage{
ggSC.link(outcome)
}
\arguments{
\item{outcome}{a table generated from SC function}
}
\value{
a figure of estimated sample completeness with order q
}
\description{
\code{ggSC.link} The figure for estimation of Sample Completeness with order q
}
\examples{
data(Norfolk)
output = SC.link(Norfolk)
ggSC.link(output)
}
