% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNEXTlink.R
\name{ggSpec}
\alias{ggSpec}
\title{ggplot for Evenness
\code{ggNetEven} The figure for estimation of Evenness with order q\cr}
\usage{
ggSpec(output)
}
\arguments{
\item{output}{a table generated from Evenness function\cr}
}
\value{
a figure of estimated sample completeness with order q\cr
}
\description{
ggplot for Evenness
\code{ggNetEven} The figure for estimation of Evenness with order q\cr
}
\examples{
\dontrun{
data(Norfolk)
Est <- Specialization(x = Norfolk, datatype = "abundance", q = c(0,1,2), nboot = 30, method = "Estimated")
Emp <- Specialization(x = Norfolk, datatype = "abundance", q = c(0,1,2), nboot = 30, method = "Empirical")
Est
Emp
ggSpec(Est)
ggSpec(Est)
}
}
\references{
Chao,A.and Ricotta,C.(2019).Quantifying evenness and linking it to diversity, beta diversity, and similarity.
}
