% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNEXTlink.R
\name{iNEXT.link}
\alias{iNEXT.link}
\title{Interpolation (rarefaction) and extrapolation of Chao et al.’s (2021) network diversity and mean network diversity
Function \code{iNEXT.link} computes network diversity estimates for rarefied samples and extrapolated samples
along with confidence intervals and related coverage estimates based on Chao et al.’s (2021) network
diversity (ND)}
\usage{
iNEXT.link(
  x,
  diversity = "TD",
  q = c(0, 1, 2),
  datatype = "abundance",
  size = NULL,
  nT = NULL,
  endpoint = NULL,
  knots = 40,
  conf = 0.95,
  nboot = 30,
  row.tree = NULL,
  col.tree = NULL
)
}
\arguments{
\item{diversity}{a choice of three-level diversity: 'TD' = 'Taxonomic', 'PD' = 'Phylogenetic', and 'FD' = 'Functional' under certain threshold.}

\item{q}{a nonnegative value or sequence specifying the diversity order. Default is \code{c(0,1,2)}.}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}),
or species-by-site raw incidence matrix (\code{datatype = "incidence_raw"}). Default is \code{"abundance"}.}

\item{size}{a sequence of positive integers specifying the sample sizes for which PD or meanPD estimates will be calculated.
If \code{NULL}, then estimates will be calculated for those sample sizes determined by the specified/default \code{endpoint}
and \code{knots}.}

\item{nT}{needed only when \code{datatype = "incidence_raw"}, a sequence of named nonnegative integers specifying the number of sampling units in each assemblage.
If \code{names(nT) = NULL}, then assemblage are automatically named as "assemblage1", "assemblage2",..., etc. Ignored if \code{datatype = "abundance"}.}

\item{endpoint}{a positive integer specifying the endpoint for the rarefaction and extrapolation range.
If \code{NULL}, then \code{endpoint} = double of the reference sample size in each assemblage. It is ignored if \code{size} is given.}

\item{knots}{a positive integer specifying the number of equally-spaced knots between 1 and the \code{endpoint}. Default is 40.}

\item{conf}{a positive number < 1 specifying the level of confidence interval. Default is 0.95.}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing sampling uncertainty and constructing confidence intervals.
Enter 0 to skip the bootstrap procedures. Default is 50.}

\item{row.tree}{phylogenetic tree of row assemblage in interaction matrix.}

\item{col.tree}{phylogenetic tree of column assemblage in interaction matrix}

\item{data}{a matrix/data.frame of species
abundances (for abundance data) or species-by-site incidence raw matrix/data.frame (for incidence data).\cr
Abundance data: a species-by-site matrix/data.frame of species abundances. The row (species) names of
data must match the species names in the phylogenetic tree and thus cannot be missing.\cr
Incidence raw data: species-by-site raw incidence matrix/data.frame. When there are N assemblages
and thus N matrices, users must first merge the N matrices by species identity to obtain a large
merged incidence matrix, where the rows of the matrix refer to all species presented in the merged
data. The row (species) names of data must match the species names in the phylogenetic tree and
thus cannot be missing.}

\item{type}{desired diversity type: \code{type = "PD"} for Chao et al. (2010) phylogenetic diversity
and \code{type = "meanPD"} for mean phylogenetic diversity (phylogenetic Hill number). Default is \code{"PD"}.}
}
\value{
\itemize{
 \item{\code{$DataInfo}: A dataframe summarizing data information}
 \item{\code{$iNextEst}: coverage-based diversity estimates along with confidence intervals}
 (if \code{nboot > 0}) for showing diversity estimates for rarefied and extrapolated samples along with related statistics;
 \item{\code{$AsyEst}: for
showing asymptotic diversity estimates along with related statistics.}
}
}
\description{
Interpolation (rarefaction) and extrapolation of Chao et al.’s (2021) network diversity and mean network diversity
Function \code{iNEXT.link} computes network diversity estimates for rarefied samples and extrapolated samples
along with confidence intervals and related coverage estimates based on Chao et al.’s (2021) network
diversity (ND)
}
\examples{
\dontrun{
data(puerto.rico)
iNEXT.link(puerto.rico$data, diversity = 'TD', datatype="abundance")
iNEXT.link(puerto.rico$data, diversity = 'PD', datatype="abundance", row.tree = puerto.rico$row.tree, col.tree = puerto.rico$col.tree)
}
}
\references{
Chao, A., Chiu C.-H. and Jost, L. (2010). Phylogenetic diversity measures based on Hill numbers. \emph{Philosophical Transactions of the Royal Society B.}, 365, 3599-3609. \cr\cr
Chao, A., Chiu, C.-H., Hsieh, T. C., Davis, T., Nipperess, D., and Faith, D. (2015). Rarefaction and extrapolation of phylogenetic diversity. \emph{Methods in Ecology and Evolution}, 6, 380-388.\cr\cr
Chao, A., Chiu C.-H. and Jost L. (2016). Phylogenetic diversity measures and their decomposition: a framework based on Hill numbers. pp. 141-172 in Pellens R. and Grandcolas P. (eds)
\emph{Biodiversity Conservation and Phylogenetic Systematics: Preserving our Evolutionary Heritage in an Extinction Crisis}, Springer. \cr\cr
Hsieh, T. C. and Chao, A. (2017). Rarefaction and extrapolation: making fair comparison of abundance-sensitive phylogenetic diversity among multiple assemblages. \emph{Systematic Biology}, 66, 100-111.
}
