% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNEXTlink.R
\name{NetSC}
\alias{NetSC}
\title{Sample Completeness main function}
\usage{
NetSC(x, q = seq(0, 2, 0.2), datatype = "abundance", nboot = 30, conf = 0.95)
}
\arguments{
\item{x}{a matrix/data.frame/list/vector of abundances-based/incidences-based species data.\cr}

\item{q}{a integer vector for the order of Hill number\cr}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}),
sampling-unit-based incidence frequencies data (\code{datatype = "incidence_freq"}) or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}).\cr}

\item{nboot}{an integer specifying the number of bootstrap replications, default is 30.\cr}

\item{conf}{positive number < 1 specifying the level of confidence interval, default is 0.95.\cr\cr}
}
\value{
a matrix of estimated sample completeness with order q: \cr\cr
}
\description{
\code{NetSC} Estimation of Sample Completeness with order q
}
\examples{
data(Norfolk)
output = NetSC(Norfolk)
ggNetSC(output)

}
\references{
Chao,A.,Y.Kubota,D.Zelený,C.-H.Chiu.
Quantifying sample completeness and comparing diversities among assemblages.
}
