% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNEXTlink.R
\name{Obs3D.link}
\alias{Obs3D.link}
\title{Empirical diversity q profile}
\usage{
Obs3D.link(
  data,
  class = "TD",
  q = seq(0, 2, 0.2),
  datatype = "abundance",
  nboot = 50,
  conf = 0.95,
  col.tree = NULL,
  row.tree = NULL
)
}
\arguments{
\item{outcome}{the outcome of the functions \code{ObsND} .\cr}
}
\value{
a table of Asymptotic network diversity q profile
}
\description{
\code{Obs3D.link} The estimated and empirical diversity of order q
}
\examples{
\dontrun{
## Example for abundance-based data
## Ex.1
data(Norfolk)
out1 <- Obs3D.link(Norfolk, class = 'TD', datatype = "abundance", nboot = 30)
ggObs3D.link(out1)
## Ex.2
data(puerto.rico)
out2 <- Obs3D.link(data = puerto.rico$data, class = 'PD', datatype = "abundance",
ggObs3D.link(out2)
}
}
