% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{ObsAsy.link}
\alias{ObsAsy.link}
\title{Asymptotic and Observed diversity q profile}
\usage{
ObsAsy.link(
  data,
  diversity = "TD",
  q = seq(0, 2, 0.2),
  nboot = 30,
  conf = 0.95,
  method = c("Asymptotic", "Observed"),
  row.tree = NULL,
  col.tree = NULL,
  PDtype = "meanPD",
  row.distM = NULL,
  col.distM = NULL,
  FDtype = "AUC",
  FDtau = NULL
)
}
\arguments{
\item{data}{a \code{list} of \code{data.frames}, each \code{data.frames} represents col.species-by-row.species abundance matrix.}

\item{diversity}{selection of diversity type: \code{'TD'} = Taxonomic diversity, \code{'PD'} = Phylogenetic diversity, and \code{'FD'} = Functional diversity.}

\item{q}{a numerical vector specifying the diversity orders. Default is \code{seq(0, 2, 0.2)}.}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing sampling uncertainty and constructing confidence intervals. Bootstrap replications are generally time consuming. Enter 0 to skip the bootstrap procedures. Default is \code{30}.}

\item{conf}{a positive number < 1 specifying the level of confidence interval. Default is \code{0.95}.}

\item{method}{asymptotic or Observed}

\item{row.tree}{(required only when \code{diversity = "PD"}), a phylogenetic tree of row assemblage in the pooled network row assemblage.}

\item{col.tree}{(required only when \code{diversity = "PD"}), a phylogenetic tree of column assemblage in the pooled network column assemblage.}

\item{PDtype}{(required only when \code{diversity = "PD"}), select PD type: \code{PDtype = "PD"}(effective total branch length) or \code{PDtype = "meanPD"}(effective number of equally divergent lineages).Default is \code{"meanPD"}.}

\item{row.distM}{(required only when \code{diversity = "FD"}), a species pairwise distance matrix for all species of row assemblage in the pooled network row assemblage.}

\item{col.distM}{(required only when \code{diversity = "FD"}), a species pairwise distance matrix for all species of column assemblage in the pooled network column assemblage.}

\item{FDtype}{(required only when \code{diversity = "FD"}), select FD type: \code{FDtype = "tau_values"} for FD under specified threshold values, or \code{FDtype = "AUC"} (area under the curve of tau-profile) for an overall FD which integrates all threshold values between zero and one. Default is \code{"AUC"}.}

\item{FDtau}{(required only when \code{diversity = "FD"} and \code{FDtype = "tau_values"}), a numerical vector between 0 and 1 specifying tau values (threshold levels). If \code{NULL} (default), then threshold is set to be the mean distance between any two individuals randomly selected from the pooled assemblage (i.e., quadratic entropy).}
}
\value{
a table of diversity table including the following arguments.
\item{Order.q}{the diversity order of q.}
\item{qTD, qPD, qFD}{the estimated asymptotic diversity or observed diversity of order q.}
\item{s.e.}{standard error of diversity.}
\item{qTD.LCL, qPD.LCL, qFD.LCL and qTD.UCL, qPD.UCL, qFD.UCL}{the bootstrap lower and upper confidence limits for the diversity of order q at the specified level (with a default value of 0.95).}
\item{Assemblage (or Network)}{the network name.}
\item{Method}{"Asymptotic" means asymptotic diversity and "Observed" means observed diversity.}
\item{Reftime}{the reference times for PD.}
\item{Type}{"PD" (effective total branch length) or "meanPD" (effective number of equally divergent lineages) for PD.}
\item{Tau}{the threshold of functional distinctiveness between any two species for FD (under FDtype = tau_values).}
}
\description{
\code{ObsAsy.link}: The asymptotic (or observed) diversity of order q
}
\examples{
## Taxonomic diversity
data(beetles)
output1 = ObsAsy.link(data = beetles, diversity = 'TD', q = seq(0, 2, 0.2))
output1


## Phylogenetic diversity
data(beetles)
data(beetles_col_tree)
output2 = ObsAsy.link(data = beetles, diversity = 'PD', q = seq(0, 2, 0.2), col.tree = beetles_col_tree)
output2


## Functional diversity under single threshold
data(beetles)
data(beetles_col_distM)
output3 = ObsAsy.link(data = beetles, diversity = 'FD', q = seq(0, 2, 0.2), col.distM = beetles_col_distM, FDtype = "tau_values")
output3


## Functional diversity with thresholds integrating from 0 to 1
data(beetles)
data(beetles_col_distM)
output4 = ObsAsy.link(data = beetles, diversity = 'FD', q = seq(0, 2, 0.25),
                      col.distM = beetles_col_distM, FDtype = "AUC", nboot = 0)
output4
}
