% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNEXTlink.R
\name{ObsND}
\alias{ObsND}
\title{Empirical diversity q profile}
\usage{
ObsND(
  data,
  q = seq(0, 2, 0.2),
  datatype = "abundance",
  nboot = 50,
  conf = 0.95
)
}
\arguments{
\item{outcome}{the outcome of the functions \code{ObsND} .\cr}
}
\value{
a table of Asymptotic network diversity q profile
}
\description{
\code{ObsND} The estimated and empirical diversity of order q
}
\examples{
## Type (1) example for abundance-based data
## Ex.1
data(spider)
out1 <- ObsND(spider, class = 'TD', datatype = "abundance")
ggObsPND(out1)

## Type (2) example for incidence-based data

## Ex.2
data(ant)
out2 <- ObsND(ant, class = 'TD', datatype = "incidence_freq", nboot = 0)
ggObsPND(out2)

}
