% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNEXTlink.R
\name{ObsPND}
\alias{ObsPND}
\title{Empirical diversity q profile}
\usage{
ObsPND(
  data,
  q = seq(0, 2, 0.2),
  datatype = "abundance",
  nboot = 50,
  conf = 0.95,
  row.tree = NULL,
  col.tree = NULL
)
}
\arguments{
\item{outcome}{the outcome of the functions \code{ObsND} .\cr}
}
\value{
a table of Asymptotic network diversity q profile
}
\description{
\code{ObsND} The estimated and empirical diversity of order q
}
\examples{
\dontrun{
## Example for abundance-based data
data(puerto.rico$data)
out1 <- ObsPND(puerto.rico$data, datatype = "abundance", row.tree = row.tree,col.tree = col.tree)
ggObsPND(out1)
}
}
