% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{Spec.link.ObsAsy}
\alias{Spec.link.ObsAsy}
\title{Asymptotic estimation (or observed) of specialization with order q}
\usage{
Spec.link.ObsAsy(
  data,
  q = seq(0, 2, 0.2),
  method = "Asymptotic",
  nboot = 30,
  conf = 0.95,
  E.class = c(1:5),
  level = "weighted"
)
}
\arguments{
\item{data}{a \code{list} of \code{data.frames}, each \code{data.frames} represents col.species-by-row.species abundance matrix.}

\item{q}{a numerical vector specifying the diversity orders. Default is \code{seq(0, 2, 0.2)}.}

\item{method}{a binary calculation method with \code{"Asymptotic"} or \code{"Observed"}.}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing
sampling uncertainty and constructing confidence intervals. Bootstrap replications are generally time consuming. Enter 0 to skip the bootstrap procedures. Default is \code{30}.}

\item{conf}{a positive number < 1 specifying the level of confidence interval. Default is \code{0.95}.}

\item{E.class}{an integer vector between 1 to 5.}

\item{level}{the type of specialization measure: Choose "weighted" for weighted species-level specialization (default), or "network" for network-level specialization.}
}
\value{
A list of several tables containing estimated (or observed) evenness with order q.\cr
        Each tables represents a class of specialization.
        \item{Order.q}{the network diversity order of q.}
        \item{Specialization}{the specialization of order q.}
        \item{s.e.}{standard error of evenness.}
        \item{Spec.LCL, Spec.UCL}{the bootstrap lower and upper confidence limits for the evenness of order q at the specified level (with a default value of \code{0.95}).}
        \item{Method}{\code{"Asymptotic"} or \code{"Observed"}.}
        \item{Dataset}{the dataset name.}
        \item{Measure}{specialization class.}
        \item{level}{type of specialization measure.}
}
\description{
Asymptotic estimation (or observed) of specialization with order q
}
\examples{
data(beetles)
output_spec_ObsAsy = Spec.link.ObsAsy(beetles)
output_spec_ObsAsy
}
