\name{beetles_row_distM}
\alias{beetles_row_distM}
\docType{data}
\title{
Species pairwise distance matrix for species in the dataset \code{beetles} and \code{beetles_plotA}.
}
\description{
This dataset includes species pairwise symmetric distance matrix for 238 species listed in the 
  \code{beetles} and \code{beetles_plotA} dataset. Each element in the matrix represents a Gower distance between two interaction computed from species traits.
}
\usage{data("beetles_row_distM")}
\format{
  A data frame with 264 species on the following 264 species.
  \describe{
    \item{\code{Tachyta_nana}}{a numeric vector}
    \item{\code{Plegaderus_vulneratus}}{a numeric vector}
    \item{\code{Plegaderus_caesus}}{a numeric vector}
    ....... \cr
  }
}
\references{
Vogel, S., et al. (2020). Optimizing enrichment of deadwood for biodiversity by varying sun   exposure and tree species: an experimental approach. Journal of Applied Ecology 57(10): 2075-2085.
}

\keyword{datasets}
