% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{Completeness.link}
\alias{Completeness.link}
\title{Sample Completeness main function}
\usage{
Completeness.link(data, q = seq(0, 2, 0.2), nboot = 30, conf = 0.95)
}
\arguments{
\item{data}{a \code{list} of \code{data.frames}, each \code{data.frames} represents col.species-by-row.species abundance matrix.}

\item{q}{q a numerical vector specifying the diversity orders. Default is \code{seq(0, 2, 0.2)}.}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing
sampling uncertainty and constructing confidence intervals. Bootstrap replications are generally time consuming. Enter 0 to skip the bootstrap procedures. Default is \code{30}.}

\item{conf}{a positive number < \code{1} specifying the level of confidence interval. Default is \code{0.95}.}
}
\value{
a matrix of estimated sample completeness with order q: 
        \item{Order.q}{the diversity order of q.}
        \item{Estimate.SC}{the estimated (or observed) sample completeness of order q.}
        \item{s.e.}{standard error of sample completeness.}
        \item{SC.LCL, SC.UCL}{the bootstrap lower and upper confidence limits for the sample completeness of order q at the specified level (with a default value of \code{0.95}).}
        \item{Assemblage}{the assemblage name.}
}
\description{
\code{Completeness.link} Estimation of Sample Completeness with order q
}
\examples{
data(beetles)
output = Completeness.link(beetles)
output

}
\references{
Chao, A., Y. Kubota, D. Zelen??, C.-H. Chiu, C.-F. Li, B. Kusumoto, M. Yasuhara, S. Thorn, C.-L. Wei, M. J. Costello, and R. K. Colwell (2020). Quantifying sample completeness and comparing diversities among assemblages. Ecological Research, 35, 292-314.
}
