% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{DataInfo.link}
\alias{DataInfo.link}
\title{Exhibit basic data information}
\usage{
DataInfo.link(
  data,
  diversity = "TD",
  row.tree = NULL,
  col.tree = NULL,
  row.distM = NULL,
  col.distM = NULL
)
}
\arguments{
\item{data}{a \code{list} of \code{data.frames}, each \code{data.frames} represents col.species-by-row.species abundance matrix.}

\item{diversity}{selection of diversity type: \code{'TD'} = 'Taxonomic diversity', \code{'PD'} = 'Phylogenetic diversity', and \code{'FD'} = 'Functional diversity'.}

\item{row.tree}{(required only when \code{diversity = "PD"}), a phylogenetic tree of row assemblage in the pooled network row assemblage.}

\item{col.tree}{(required only when \code{diversity = "PD"}), a phylogenetic tree of column assemblage in the pooled network column assemblage.}

\item{row.distM}{(required only when \code{diversity = "FD"}), a species pairwise distance matrix for all species of row assemblage in the pooled network row assemblage.}

\item{col.distM}{(required only when \code{diversity = "FD"}), a species pairwise distance matrix for all species of column assemblage in the pooled network column assemblage.}
}
\value{
a data.frame of basic data information including network name (\code{Networks}), sample size (\code{n}), observed species richness in row assemblage (\code{S.obs(row)}), observed species richness in column assemblage (\code{S.obs(col)}), the number of interactions (\code{Link.obs}), link percentage(\code{Connectance}), sample coverage estimate (\code{Coverage}).\cr\cr
Besides, show the first ten species abundance (or incidence) frequency counts in the reference sample in TD. (\code{f1}-\code{f10})\cr\cr
In PD, show the the observed total branch length in the phylogenetic tree (\code{PD.obs}), the number of singletons and doubletons in the node/branch set (\code{f1*}-\code{f2*}), the total branch length of those singletons/doubletons in the node/branch set (\code{g1}-\code{g2}), mean reference time (\code{mean_T}).\cr\cr
In FD (\code{FDtype = "tau_values"}), show the number of singletons and doubletons in the data (\code{f1}-\code{f2}), the number of singletons and doubletons in the functional group (\code{a1*}-\code{a2*}), the threshold of functional distinctiveness between any two species (\code{threshold}).\cr\cr
}
\description{
\code{DataInfo.link}: exhibits basic data information
}
\examples{
#' ## Taxonomic diversity
data(beetles)
DataInfo.link(data = beetles, diversity = 'TD')


## Phylogenetic diversity
data(beetles)
data(beetles_col_tree)
DataInfo.link(data = beetles, diversity = 'PD', col.tree = beetles_col_tree)


## Functional diversity
data(beetles)
data(beetles_col_distM)
DataInfo.link(data = beetles, diversity = 'FD', col.distM = beetles_col_distM)
}
