% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{estimateD.link}
\alias{estimateD.link}
\title{Compute species diversity with a particular of sample size/coverage}
\usage{
estimateD.link(
  data,
  diversity = "TD",
  q = c(0, 1, 2),
  base = "coverage",
  level = NULL,
  nboot = 50,
  conf = 0.95,
  row.tree = NULL,
  col.tree = NULL,
  PDtype = "meanPD",
  row.distM = NULL,
  col.distM = NULL,
  FDtype = "AUC",
  FDtau = NULL
)
}
\arguments{
\item{data}{a \code{matrix}, \code{data.frame} (species by assemblages), or \code{list} of species abundance/incidence raw data.\cr}

\item{diversity}{a choice of three-level diversity: 'TD' = 'Taxonomic', 'PD' = 'Phylogenetic', and 'FD' = 'Functional' under certain threshold.}

\item{q}{a numerical vector of the order of Hill number. Default is \code{seq(0, 2, 0.2)}.}

\item{base}{comparison base: sample-size-based (\code{base="size"}) or coverage-based \cr (\code{base="coverage"}).}

\item{level}{a sequence specifying the particular sample sizes or sample coverages(between 0 and 1).
If \code{base="size"} and \code{level=NULL}, then this function computes the diversity estimates for the minimum sample size among all sites extrapolated to double reference sizes.
If \code{base="coverage"} and \code{level=NULL}, then this function computes the diversity estimates for the minimum sample coverage among all sites extrapolated to double reference sizes.}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing
sampling uncertainty and constructing confidence intervals. Enter 0 to skip the bootstrap procedures. Default is \code{50}.}

\item{conf}{a positive number < 1 specifying the level of confidence interval, default is \code{0.95}.}

\item{row.tree}{phylogenetic tree of row assemblage in interaction matrix.}

\item{col.tree}{phylogenetic tree of column assemblage in interaction matrix.}

\item{PDtype}{(required only when \code{diversity = "PD"}), select PD type: \code{PDtype = "PD"}(effective total branch length) or
\code{PDtype = "meanPD"}(effective number of equally divergent lineages).Default is \code{"meanPD"}.}

\item{row.distM}{(required only when \code{diversity = "FD"}), a row species pairwise distance matrix for all row species of row assemblage in interaction matrix.}

\item{col.distM}{(required only when \code{diversity = "FD"}), a column species pairwise distance matrix for all column species of column assemblage in interaction matrix.}

\item{FDtype}{(required only when \code{diversity = "FD"}), select FD type: \code{FDtype = "tau_values"} for FD under specified threshold values, or \code{FDtype = "AUC"} (area under the curve of tau-profile) for an overall FD which integrates all threshold values between zero and one. Default is \code{"AUC"}.}

\item{FDtau}{(required only when \code{diversity = "FD"} and \code{FDtype = "tau_values"}), a numerical vector between 0 and 1 specifying tau values (threshold levels). If \code{NULL} (default), then threshold is set to be the mean distance between any two individuals randomly selected from the pooled assemblage (i.e., quadratic entropy).}
}
\value{
a \code{data.frame} of diversity table including the following arguments:
\item{Dataset}{the name of datasets.}
\item{Order.q}{the diversity order of q.}
\item{SC}{the target standardized coverage value.}
\item{m}{the corresponding sample size for the standardized completeness.}
\item{Method}{Rarefaction, Observed, or Extrapolation, depending on whether the target coverage is less than, equal to, or greater than the coverage of the reference sample.}
\item{qiTD, qiPD, qiFD}{the estimated network diversity of order q for the target coverage value The estimate for complete coverage (or \code{level = 1}) represents the estimated asymptotic diversity.} 
\item{s.e.}{standard error of network diversity estimate.}
\item{qiTD.LCL, qiPD.LCL, qiFD.LCL and qiTD.UCL, qiPD.UCL, qiFD.UCL}{the bootstrap lower and upper confidence limits for the network diversity of order q at the specified level (with a default value of \code{0.95}).}
\item{Reftime}{reference times for PD.}
\item{Type}{\code{"PD"} (phylogenetic network diversity of effective total branch length), \code{"meanPD"} (effective number of equally divergent lineages).}
}
\description{
\code{estimateD.link} computes species diversity (Hill numbers with q = 0, 1 and 2) with a particular user-specified level of sample size or sample coverage.
}
\examples{
\donttest{
# Taxonomic network diversity for interaction data with two target coverages (93\% and 97\%)
data(beetles_plotA)
output_est_qiTD <- estimateD.link(beetles_plotA, diversity = 'TD', q = c(0,1,2),
                                  base = "coverage", level = c(0.93, 0.97))
output_est_qiTD

# Phylogenetic network diversity for interaction data with two target sizes (1500 and 3000)
data(beetles_plotA)
data(beetles_col_tree)
output_est_qiPD <- estimateD.link(beetles_plotA, diversity = 'PD', 
                                  base = "size", level = c(1500, 3000), col.tree = beetles_col_tree)
output_est_qiPD

## Functional network diversity for interaction data with two target coverages (93\% and 97\%)
data(beetles_plotA)
data(beetles_col_distM)
output_est_qiFD = estimateD.link(data = beetles_plotA, diversity = 'FD', q = c(0, 1, 2),
                                 base = "coverage", level = c(0.93, 0.97), nboot = 10,
                                 col.distM = beetles_col_distM, FDtype = "AUC")
output_est_qiFD


}
}
