% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{estimateD.link}
\alias{estimateD.link}
\title{Compute species diversity with a particular of sample size/coverage}
\usage{
estimateD.link(
  data,
  diversity = "TD",
  q = c(0, 1, 2),
  datatype = "abundance",
  base = "size",
  level = NULL,
  nboot = 50,
  conf = 0.95,
  row.tree = NULL,
  col.tree = NULL,
  row.distM = NULL,
  col.distM = NULL,
  FDtype = "AUC",
  FDtau = NULL
)
}
\arguments{
\item{data}{a \code{matrix}, \code{data.frame} (species by assemblages), or \code{list} of species abundance/incidence raw data.\cr}

\item{diversity}{a choice of three-level diversity: 'TD' = 'Taxonomic', 'PD' = 'Phylogenetic', and 'FD' = 'Functional' under certain threshold. Besides,'AUC' is the fourth choice which
integrates several threshold functional diversity to get diversity.}

\item{q}{a numerical vector of the order of Hill number.}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}) or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}).}

\item{base}{comparison base: sample-size-based (\code{base="size"}) or coverage-based \cr (\code{base="coverage"}).}

\item{level}{a sequence specifying the particular sample sizes or sample coverages(between 0 and 1).
If \code{base="size"} and \code{level=NULL}, then this function computes the diversity estimates for the minimum sample size among all sites extrapolated to double reference sizes.
If \code{base="coverage"} and \code{level=NULL}, then this function computes the diversity estimates for the minimum sample coverage among all sites extrapolated to double reference sizes.}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing
sampling uncertainty and constructing confidence intervals. Enter 0 to skip the bootstrap procedures. Default is 50}

\item{conf}{a positive number < 1 specifying the level of confidence interval, default is 0.95.}

\item{row.tree}{phylogenetic tree of row assemblage in interaction matrix.}

\item{col.tree}{phylogenetic tree of column assemblage in interaction matrix.}

\item{row.distM}{(required only when diversity = "FD"), a row species pairwise distance matrix for all row species of row assemblage in interaction matrix.}

\item{col.distM}{(required only when diversity = "FD"), a column species pairwise distance matrix for all column species of column assemblage in interaction matrix.}

\item{FDtype}{(required only when diversity = "FD"), select FD type: FDtype = "tau_values" for FD under specified threshold values, or FDtype = "AUC" (area under the curve of tau-profile) for an overall FD which integrates all threshold values between zero and one. Default is "AUC".}

\item{FDtau}{(required only when diversity = "FD" and FDtype = "tau_values"), a numerical vector between 0 and 1 specifying tau values (threshold levels). If NULL (default), then threshold is set to be the mean distance between any two individuals randomly selected from the pooled assemblage (i.e., quadratic entropy).}
}
\value{
a data.frame of species diversity table including the sample size, sample coverage, method (rarefaction or extrapolation), and diversity estimates with q = 0, 1, and 2 for the user-specified sample size or sample coverage.
}
\description{
\code{estimateD.link} computes species diversity (Hill numbers with q = 0, 1 and 2) with a particular user-specified level of sample size or sample coverage.
}
\examples{
\dontrun{
data(beetles)
out1 <- estimateD.link(beetles, diversity = 'TD', datatype = "abundance",
                       base = "coverage", level = 0.7, nboot = 30)
out2 <- estimateD.link(beetles, diversity = 'TD', datatype = "abundance",
                       base = "size", level = NULL, nboot = 30)
}
}
