% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNEXTlink.R
\name{estimateND}
\alias{estimateND}
\title{Compute species diversity with a particular of sample size/coverage}
\usage{
estimateND(
  dat,
  q = c(0, 1, 2),
  datatype = "abundance",
  base = "size",
  level = NULL,
  nboot = 50,
  conf = 0.95
)
}
\arguments{
\item{outcome}{the outcome of the functions \code{ObsND} .\cr}
}
\value{
a data.frame of species diversity table including the sample size, sample coverage, method (rarefaction or extrapolation), and diversity estimates with q = 0, 1, and 2 for the user-specified sample size or sample coverage.
}
\description{
\code{estimateND} computes species diversity (Hill numbers with q = 0, 1 and 2) with a particular user-specified level of sample size or sample coverage.
}
\examples{
## Not run:
data(spider)
out <- estimateD(spider, q = c(0,1,2), datatype = "abundance", base="size")
out <- estimateD(spider, q = c(0,1,2), datatype = "abundance", base="coverage")

}
