% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNEXTlink.R
\name{ggAsyPND}
\alias{ggAsyPND}
\title{ggplot for Asymptotic diversity}
\usage{
ggAsyPND(outcome, text_size = 10)
}
\arguments{
\item{outcome}{the outcome of the functions \code{AsyD} .\cr}
}
\value{
a figure of estimated sample completeness with order q\cr\cr
}
\description{
\code{ggAsyPND} Plots q-profile based on the outcome of \code{AsyD} using the ggplot2 package.\cr
It will only show the confidence interval of 'Estimated'.
}
\examples{
\dontrun{
data(puer)
out1 <- AsyPND(puer)
ggAsyPND(out1)
}

}
