% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{ggCompleteness.link}
\alias{ggCompleteness.link}
\title{ggplot for Sample Completeness}
\usage{
ggCompleteness.link(output)
}
\arguments{
\item{output}{a table generated from Completeness.link function}
}
\value{
a figure of estimated sample completeness with order q
}
\description{
\code{ggCompleteness.link} The figure for estimation of Sample Completeness with order q
}
\examples{
# Plot the completeness curve
data(beetles_plotA)
output_com = Completeness.link(beetles_plotA)
ggCompleteness.link(output_com)

}
\references{
Chao, A., Y. Kubota, D. Zeleny, C.-H. Chiu, C.-F. Li, B. Kusumoto, M. Yasuhara, S. Thorn, C.-L. Wei, M. J. Costello, and R. K. Colwell (2020). Quantifying sample completeness and comparing diversities among assemblages. Ecological Research, 35, 292-314.
}
