% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{ggiNEXT.link}
\alias{ggiNEXT.link}
\title{ggplot2 extension for outcome from \code{iNEXT.link}}
\usage{
ggiNEXT.link(
  outcome,
  type = c(1, 2, 3),
  facet.var = "Assemblage",
  color.var = "Order.q"
)
}
\arguments{
\item{outcome}{a list object computed by \code{\link{iNEXT.link}}.}

\item{type}{three types of plots: sample-size-based rarefaction/extrapolation curve (\code{type = 1});
sample coverage curve (\code{type = 2}); coverage-based rarefaction/extrapolation curve (\code{type = 3}).}

\item{facet.var}{create a separate plot for each value of a specified variable:
no separation \cr (\code{facet.var="None"});
a separate plot for each diversity order (\code{facet.var="Order.q"});
a separate plot for each assemblage (\code{facet.var="Assemblage"});
a separate plot for each combination of order x assemblage (\code{facet.var="Both"}).}

\item{color.var}{create curves in different colors for values of a specified variable:
all curves are in the same color (\code{color.var="None"});
use different colors for diversity orders (\code{color.var="Order.q"});
use different colors for sites (\code{color.var="Assemblage"});
use different colors for combinations of order x assemblage (\code{color.var="Both"}).}

\item{...}{other arguments passed on to methods. Not currently used.}
}
\value{
a ggplot2 object for coverage-based or size-based rarefaction and extrapolation
}
\description{
\code{ggiNEXT.link}: the \code{\link[ggplot2]{ggplot}} extension for \code{\link{iNEXT.link}} Object to plot sample-size-based and coverage-based rarefaction/extrapolation curves along with a bridging sample completeness curve
}
\examples{
## Taxonomic diversity
data(beetles)
output1 = iNEXT.link(data = beetles, diversity = 'TD', q = c(0,1,2))
ggiNEXT.link(output1)


## Phylogenetic diversity
data(beetles)
data(beetles_col_tree)
output2 = iNEXT.link(data = beetles, diversity = 'PD', q = c(0,1,2), col.tree = beetles_col_tree)
ggiNEXT.link(output2)


## Functional diversity under single threshold
data(beetles)
data(beetles_col_distM)
output3 = iNEXT.link(data = beetles, diversity = 'FD', q = c(0,1,2), nboot = 0, col.distM = beetles_col_distM, FDtype = "tau_values")
ggiNEXT.link(output3)


## Functional diversity with thresholds integrating from 0 to 1
data(beetles)
data(beetles_col_distM)
output4 = iNEXT.link(data = beetles, diversity = 'FD', q = c(0,1,2), nboot = 0, col.distM = beetles_col_distM, FDtype = "AUC")
ggiNEXT.link(output4)
}
