% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNEXTlink.R
\name{ggiNEXTbeta.link}
\alias{ggiNEXTbeta.link}
\title{ggplot2 extension for outcome from \code{iNEXT.beta.link}}
\usage{
ggiNEXTbeta.link(
  output,
  type = c("B", "D"),
  diversity = "TD",
  scale = "free",
  main = NULL,
  transp = 0.4,
  stript.size = 11,
  text.size = 13
)
}
\arguments{
\item{output}{an list object computed by \code{\link{iNEXT.beta.link}}.}

\item{type}{three types of plots: sample-size-based rarefaction/extrapolation curve (\code{type = 1});
sample completeness curve (\code{type = 2}); coverage-based rarefaction/extrapolation curve (\code{type = 3}).}

\item{se}{a logical variable to display confidence interval around the estimated sampling curve.}

\item{facet.var}{create a separate plot for each value of a specified variable:
no separation \cr (\code{facet.var="None"});
a separate plot for each diversity order (\code{facet.var="Order.q"});
a separate plot for each assemblage (\code{facet.var="Assemblage"});
a separate plot for each combination of order x assemblage (\code{facet.var="Both"}).}

\item{color.var}{create curves in different colors for values of a specified variable:
all curves are in the same color (\code{color.var="None"});
use different colors for diversity orders (\code{color.var="Order.q"});
use different colors for sites (\code{color.var="Assemblage"});
use different colors for combinations of order x assemblage (\code{color.var="Both"}).}

\item{grey}{a logical variable to display grey and white ggplot2 theme.}

\item{...}{other arguments passed on to methods. Not currently used.}
}
\value{
a ggplot2 object
}
\description{
\code{ggiNEXTbeta.link}: the \code{\link[ggplot2]{ggplot}} extension for
\code{\link{iNEXTbeta.link}} Object to plot sample-size- and coverage-based rarefaction/extrapolation curves along with a bridging sample completeness curve
}
\examples{
\dontrun{
data(Norfolk)
beta1 = iNEXTbeta.link(networks = puerto.rico$data, level = seq(0.5, 1, 0.5), datatype='abundance',q = c(0, 1, 2),
                       diversity = 'TD', nboot = 10, conf = 0.95)
beta2 = iNEXTbeta.link(networks = puerto.rico$data\%>\%lapply(function(x) round(x/10)), level = seq(0.5, 1, 0.5), datatype='abundance',q = c(0, 1, 2),
                       diversity = 'PD', nboot = 10, conf = 0.95,
                       row.tree = puerto.rico$row.tree, col.tree = puerto.rico$col.tree)

ggiNEXTbeta.link(beta2,diversity = 'TD', type = 'B')
ggiNEXTbeta.link(beta2,diversity = 'TD', type = 'D')
ggiNEXTbeta.link(beta2,diversity = 'PD', type = 'B')
ggiNEXTbeta.link(beta2,diversity = 'PD', type = 'D')
}
}
