% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{ggiNEXTBeta.link}
\alias{ggiNEXTBeta.link}
\title{ggplot2 extension for outcome from \code{iNEXTBeta.link}}
\usage{
ggiNEXTBeta.link(outcome, type = c("B", "D"), scale = "free")
}
\arguments{
\item{outcome}{the outcome from \code{"iNEXTBeta.link"}}

\item{type}{selection of plot type : \code{type = 'B'} for plotting the gamma, alpha, and beta diversity ;
\code{type = 'D'} for plotting 4 turnover dissimilarities.}

\item{scale}{Are scales shared across all facets (\code{"fixed"}), or do they vary across rows (\code{"free_x"}), columns (\code{"free_y"}), or both rows and columns (\code{"free"})? Default is \code{"free"}.}
}
\value{
a figure for Beta diversity or dissimilarity diversity.
}
\description{
\code{ggiNEXTBeta.link}: ggplot for Interpolation and extrapolation of Beta diversity with order q
}
\examples{
## Taxonomic diversity
data(beetles)
output1 = iNEXTBeta.link(data = beetles, diversity = 'TD', level = seq(0.5, 0.9, 0.4), q = c(0, 1, 2))
ggiNEXTBeta.link(output1, type = 'B')
ggiNEXTBeta.link(output1, type = 'D')

## Phylogenetic diversity
data(beetles)
data(beetles_col_tree)
output2 = iNEXTBeta.link(data = beetles, diversity = 'PD', level = seq(0.5, 0.9, 0.4), q = c(0, 1, 2), col.tree = beetles_col_tree)
ggiNEXTBeta.link(output2, type = 'B')
ggiNEXTBeta.link(output2, type = 'D')


## Functional diversity under single threshold
data(beetles)
data(beetles_col_distM)
output3 = iNEXTBeta.link(data = beetles, diversity = 'FD', level = seq(0.5, 0.9, 0.4), q = c(0, 1, 2), col.distM = beetles_col_distM, FDtype = "tau_values")
ggiNEXTBeta.link(output3, type = 'B')
ggiNEXTBeta.link(output3, type = 'D')


## Functional diversity with thresholds integrating from 0 to 1
data(beetles)
data(beetles_col_distM)
output4 = iNEXTBeta.link(data = beetles, diversity = 'FD', level = seq(0.5, 0.9, 0.4), q = c(0, 1, 2), col.distM = beetles_col_distM, FDtype = "AUC")
ggiNEXTBeta.link(output4, type = 'B')
ggiNEXTBeta.link(output4, type = 'D')
}
