% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{ggiNEXTbeta.link}
\alias{ggiNEXTbeta.link}
\title{ggplot2 extension for outcome from \code{iNEXTbeta.link}}
\usage{
ggiNEXTbeta.link(outcome, type = c("B", "D"))
}
\arguments{
\item{outcome}{the outcome from \code{"iNEXTbeta.link"}}

\item{type}{selection of plot type : \code{type = 'B'} for plotting the gamma, alpha, and beta diversity ;
\code{type = 'D'} for plotting 4 turnover dissimilarities.}
}
\value{
a figure for gamma, alpha, and beta diversity or four dissimilarity measures.
}
\description{
\code{ggiNEXTbeta.link}: ggplot for Interpolation and extrapolation of beta diversity with order q
}
\examples{
## Taxonomic diversity
data(beetles)
output1 = iNEXTbeta.link(data = beetles, diversity = 'TD', level = seq(0.5, 0.9, 0.4), q = c(0, 1, 2))
ggiNEXTbeta.link(output1, type = 'B')
ggiNEXTbeta.link(output1, type = 'D')

## Phylogenetic diversity
data(beetles)
data(beetles_col_tree)
output2 = iNEXTbeta.link(data = beetles, diversity = 'PD', level = seq(0.5, 0.9, 0.4), q = c(0, 1, 2), col.tree = beetles_col_tree)
ggiNEXTbeta.link(output2, type = 'B')
ggiNEXTbeta.link(output2, type = 'D')


## Functional diversity under single threshold
data(beetles)
data(beetles_col_distM)
output3 = iNEXTbeta.link(data = beetles, diversity = 'FD', level = seq(0.5, 0.9, 0.4), q = c(0, 1, 2), col.distM = beetles_col_distM, FDtype = "tau_value")
ggiNEXTbeta.link(output3, type = 'B')
ggiNEXTbeta.link(output3, type = 'D')


## Functional diversity with thresholds integrating from 0 to 1
data(beetles)
data(beetles_col_distM)
output4 = iNEXTbeta.link(data = beetles, diversity = 'FD', level = seq(0.5, 0.9, 0.4), q = c(0, 1, 2), col.distM = beetles_col_distM, FDtype = "AUC", nboot = 0)
ggiNEXTbeta.link(output4, type = 'B')
ggiNEXTbeta.link(output4, type = 'D')
}
