% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNEXTlink.R
\name{ggSC.link}
\alias{ggSC.link}
\title{ggplot for Sample Completeness}
\usage{
ggSC.link(output)
}
\arguments{
\item{output}{a table generated from SC function}
}
\value{
a figure of estimated sample completeness with order q
}
\description{
\code{ggNetSC} The figure for estimation of Sample Completeness with order q
}
\examples{
data(Norfolk)
output = SC.link(Norfolk)
ggSC.link(output)
}
\references{
Chao,A.,Y.Kubota,D.Zelený,C.-H.Chiu.
Quantifying sample completeness and comparing diversities among assemblages.
}
