% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{ggSpec.link}
\alias{ggSpec.link}
\title{ggplot2 extension for output from \code{Spec.link.est} and \code{Spec.link.ObsAsy}.}
\usage{
ggSpec.link(output)
}
\arguments{
\item{output}{a table generated from Specialization function}
}
\value{
a figure of estimated sample completeness with order q
}
\description{
\code{ggSpec.link}: Plots q-profile based on the output of \code{Spec.link.est} and \code{Spec.link.ObsAsy} using the ggplot2 package
}
\examples{
\donttest{
data(beetles_plotA)
output_spec = Spec.link.est(beetles_plotA, nboot = 5)
ggSpec.link(output_spec)
}

}
