% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{ggSpec.link}
\alias{ggSpec.link}
\title{ggplot for Specialization ggSpec.link The figure for estimation of Specialization with order q}
\usage{
ggSpec.link(output)
}
\arguments{
\item{output}{a table generated from Specialization function}
}
\value{
a figure of estimated sample completeness with order q
}
\description{
ggplot for Specialization ggSpec.link The figure for estimation of Specialization with order q
}
\examples{
data(beetles)
output = Spec.link(beetles)
ggSpec.link(output)
}
