% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{iNEXT.link}
\alias{iNEXT.link}
\title{Interpolation (rarefaction) and extrapolation of network diversity}
\usage{
iNEXT.link(
  data,
  diversity = "TD",
  q = c(0, 1, 2),
  size = NULL,
  endpoint = NULL,
  knots = 40,
  conf = 0.95,
  nboot = 30,
  row.tree = NULL,
  col.tree = NULL,
  PDtype = "meanPD",
  row.distM = NULL,
  col.distM = NULL,
  FDtype = "AUC",
  FDtau = NULL
)
}
\arguments{
\item{data}{a \code{list} of \code{data.frames}, each \code{data.frames} represents col.species-by-row.species abundance matrix.}

\item{diversity}{selection of diversity type: \code{'TD'} = 'Taxonomic diversity', \code{'PD'} = 'Phylogenetic diversity', and \code{'FD'} = 'Functional diversity'.}

\item{q}{a numerical vector specifying the diversity orders. Default is \code{c(0,1,2)}.}

\item{size}{an integer vector of sample sizes for which diversity estimates will be computed.
If \code{NULL}, then diversity estimates will be calculated for those sample sizes determined by the specified/default \code{endpoint}
and \code{knots}.}

\item{endpoint}{an integer vector specifying the sample size that is the \code{endpoint} for rarefaction/extrapolation.
If \code{NULL}, then \code{endpoint} = double reference sample size in each assemblage.}

\item{knots}{an integer specifying the number of equally-spaced \code{knots} between size 1 and the \code{endpoint}. Default is \code{40}.}

\item{conf}{a positive number < \code{1} specifying the level of confidence interval. Default is \code{0.95}.}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing
sampling uncertainty and constructing confidence intervals. Bootstrap replications are generally time consuming. Enter 0 to skip the bootstrap procedures. Default is \code{30}.}

\item{row.tree}{(required only when \code{diversity = "PD"}), a phylogenetic tree of row assemblage in the pooled network row assemblage.}

\item{col.tree}{(required only when \code{diversity = "PD"}), a phylogenetic tree of column assemblage in the pooled network column assemblage.}

\item{PDtype}{(required only when \code{diversity = "PD"}), select PD type: \code{PDtype = "PD"}(effective total branch length) or
\code{PDtype = "meanPD"}(effective number of equally divergent lineages).Default is \code{"meanPD"}.}

\item{row.distM}{(required only when \code{diversity = "FD"}), a species pairwise distance matrix for all species of row assemblage in the pooled network row assemblage.}

\item{col.distM}{(required only when \code{diversity = "FD"}), a species pairwise distance matrix for all species of column assemblage in the pooled network column assemblage.}

\item{FDtype}{(required only when \code{diversity = "FD"}), select FD type: \code{FDtype = "tau_values"} for FD under specified threshold values, or \code{FDtype = "AUC"} (area under the curve of tau-profile) for an overall FD which integrates all threshold values between zero and one. Default is \code{"AUC"}.}

\item{FDtau}{(required only when \code{diversity = "FD"} and \code{FDtype = "tau_values"}), a numerical vector between 0 and 1 specifying tau values (threshold levels). If \code{NULL} (default), then threshold is set to be the mean distance between any two individuals randomly selected from the pooled assemblage (i.e., quadratic entropy).}
}
\value{
a list of three objects: \cr\cr
(1) \code{$TDInfo} (\code{$PDInfo}, or \code{$FDInfo}) for summarizing data information for q = 0, 1 and 2. Refer to the output of \code{DataInfo.link} for details.
(2) \code{$TDiNextEst}(\code{$PDiNextEst}, or \code{$FDiNextEst}) for showing network diversity estimates for rarefied and extrapolated samples along with related statistics. There are two data frames: \code{"$size_based"} and \code{"$coverage_based"}.
   In \code{"$size_based"}, the output includes:
   \item{Dataset}{the name of datasets.} 
   \item{Order.q}{the network diversity order of q.}
   \item{m}{the target standardized sample size.}
   \item{Method}{Rarefaction, Observed, or Extrapolation, depending on whether the size m is less than, equal to, or greater than the reference sample size.}
   \item{qiTD, qiPD, qiFD}{the estimated network diversity of order q for a sample of size m.}
   \item{qiTD.LCL, qiPD.LCL, qiFD.LCL and qiTD.UCL, qiPD.UCL, qiFD.UCL}{the bootstrap lower and upper confidence limits for the network diversity of order q at the level specified in the settings (with a default value of 0.95).}
   \item{SC}{the estimated network coverage for a sample of size m.}
   \item{SC.LCL, SC.UCL}{the bootstrap lower and upper confidence limits for the expected sample coverage at the level specified in the settings (with a default value of 0.95).}
   \item{Reftime}{the reference times for PD.}
   \item{Type}{\code{"PD"} (phylogenetic network diversity of effective total branch length),\code{"meanPD"} (effective number of equally divergent lineages) for PD.}
 Similar output is obtained for \code{"$coverage_based"}. \cr\cr
(3) \code{$TDAsyEst}(\code{$PDAsyEst}, or \code{$FDAsyEst}): for showing asymptotic diversity estimates along with related statistics:
   \item{Dataset}{the name of datasets.} 
   \item{qTD, qPD, qFD}{the network diversity order of q.}
   \item{TD_obs, PD_obs, FD_obs}{the observed network diversity.}
   \item{TD_asy, PD_asy, FD_asy}{the asymptotic network diversity estimate.}
   \item{s.e.}{standard error of asymptotic network diversity.}
   \item{qiTD.LCL, qiPD.LCL, qiFD.LCL and qiTD.UCL, qiPD.UCL, qiFD.UCL}{the bootstrap lower and upper confidence limits for asymptotic network diversity at the specified level (with a default value of 0.95).}
   \item{Reftime}{the reference times for PD.}
   \item{Type}{\code{"PD"} (phylogenetic network diversity of effective total branch length),\code{"meanPD"} (effective number of equally divergent lineages) for PD.}
}
\description{
Interpolation (rarefaction) and extrapolation of network diversity
}
\examples{
\donttest{
# Compute standardized estimates of taxonomic network diversity for interaction data with
# order q = 0, 1, 2
data(beetles_plotA)
output_qiTD = iNEXT.link(beetles_plotA, diversity = 'TD', q = c(0,1,2), nboot = 30)
output_qiTD$TDInfo   # showing basic data information.
output_qiTD$TDiNextEst   # showing diversity estimates with rarefied and extrapolated.
output_qiTD$TDAsyEst   # showing asymptotic diversity estimates.


# Compute standardized estimates of phylogenetic network diversity for interaction data with
# order q = 0, 1, 2
data(beetles_plotA)
data(beetles_row_tree)
output_qiPD = iNEXT.link(beetles_plotA, diversity = 'PD', q = c(0, 1, 2), nboot = 20,
                         row.tree = beetles_row_tree)
output_qiPD


# Compute standardized estimates of functional network diversity for interaction data with
# order q = 0, 1, 2
data(beetles_plotA)
data(beetles_row_distM)
output_qiFD = iNEXT.link(data = beetles_plotA, diversity = 'FD', q = c(0,1,2), nboot = 20, 
                         row.distM = beetles_row_distM, FDtype = "AUC")
output_qiFD
}

}
\references{
Chao, A., Henderson, P. A., Chiu, C.-H., Moyes, F., Hu, K.-H., Dornelas, M. and Magurran, A. E. (2021). Measuring temporal change in alpha diversity: a framework integrating taxonomic, phylogenetic and functional diversity and the iNEXT.3D standardization. 
\emph{Methods in Ecology and Evolution.}, 12, 1926-1940. \cr\cr
}
