% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_function.R
\name{iNEXT.link}
\alias{iNEXT.link}
\title{Interpolation (rarefaction) and extrapolation of network diversity}
\usage{
iNEXT.link(
  data,
  diversity = "TD",
  q = c(0, 1, 2),
  datatype = "abundance",
  size = NULL,
  nT = NULL,
  endpoint = NULL,
  knots = 40,
  conf = 0.95,
  nboot = 30,
  row.tree = NULL,
  col.tree = NULL,
  PDtype = "meanPD",
  row.distM = NULL,
  col.distM = NULL,
  FDtype = "AUC",
  FDtau = NULL
)
}
\arguments{
\item{data}{a \code{list} of \code{data.frames}, each \code{data.frames} represents col.species-by-row.species abundance matrix.}

\item{diversity}{selection of diversity type: \code{'TD'} = 'Taxonomic diversity', \code{'PD'} = 'Phylogenetic diversity', and \code{'FD'} = 'Functional diversity'.}

\item{q}{a numerical vector specifying the diversity orders. Default is \code{c(0,1,2)}.}

\item{size}{an integer vector of sample sizes for which diversity estimates will be computed.
If \code{NULL}, then diversity estimates will be calculated for those sample sizes determined by the specified/default \code{endpoint}
and \code{knots}.}

\item{endpoint}{an integer vector specifying the sample size that is the \code{endpoint} for rarefaction/extrapolation.
If \code{NULL}, then \code{endpoint} = double reference sample size in each assemblage.}

\item{knots}{an integer specifying the number of equally-spaced \code{knots} between size 1 and the \code{endpoint}. Default is \code{40}.}

\item{conf}{a positive number < 1 specifying the level of confidence interval. Default is \code{0.95}.}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing
sampling uncertainty and constructing confidence intervals. Bootstrap replications are generally time consuming. Enter 0 to skip the bootstrap procedures. Default is \code{30}.}

\item{row.tree}{(required only when \code{diversity = "PD"}), a phylogenetic tree of row assemblage in the pooled network row assemblage.}

\item{col.tree}{(required only when \code{diversity = "PD"}), a phylogenetic tree of column assemblage in the pooled network column assemblage.}

\item{PDtype}{(required only when \code{diversity = "PD"}), select PD type: \code{PDtype = "PD"}(effective total branch length) or
\code{PDtype = "meanPD"}(effective number of equally divergent lineages).Default is \code{"meanPD"}.}

\item{row.distM}{(required only when \code{diversity = "FD"}), a species pairwise distance matrix for all species of row assemblage in the pooled network row assemblage.}

\item{col.distM}{(required only when \code{diversity = "FD"}), a species pairwise distance matrix for all species of column assemblage in the pooled network column assemblage.}

\item{FDtype}{(required only when \code{diversity = "FD"}), select FD type: \code{FDtype = "tau_values"} for FD under specified threshold values, or \code{FDtype = "AUC"} (area under the curve of tau-profile) for an overall FD which integrates all threshold values between zero and one. Default is \code{"AUC"}.}

\item{FDtau}{(required only when \code{diversity = "FD"} and \code{FDtype = "tau_values"}), a numerical vector between 0 and 1 specifying tau values (threshold levels). If \code{NULL} (default), then threshold is set to be the mean distance between any two individuals randomly selected from the pooled assemblage (i.e., quadratic entropy).}
}
\value{
\itemize{
 \item{\code{$DataInfo}: A dataframe summarizing data information}
 \item{\code{$iNextEst}: coverage-based diversity estimates along with confidence intervals}
 for showing diversity estimates for rarefied and extrapolated samples along with related statistics;
 \item{\code{$AsyEst}: for
showing asymptotic diversity estimates along with related statistics.}
}
}
\description{
Interpolation (rarefaction) and extrapolation of network diversity
}
\examples{
## Taxonomic diversity
data(beetles)
output1 = iNEXT.link(data = beetles, diversity = 'TD', q = c(0,1,2))
output1$DataInfo # showing basic data information.
output1$iNextEst # showing diversity estimates with rarefied and extrapolated.
output1$AsyEst # showing asymptotic diversity estimates.


## Phylogenetic diversity
data(beetles)
data(beetles_col_tree)
output2 = iNEXT.link(data = beetles, diversity = 'PD', q = c(0,1,2), col.tree = beetles_col_tree)
output2


## Functional diversity under single threshold
data(beetles)
data(beetles_col_distM)
output3 = iNEXT.link(data = beetles, diversity = 'FD', q = c(0,1,2), col.distM = beetles_col_distM, FDtype = "tau_values")
output3


## Functional diversity with thresholds integrating from 0 to 1
data(beetles)
data(beetles_col_distM)
output4 = iNEXT.link(data = beetles, diversity = 'FD', q = c(0,1,2), col.distM = beetles_col_distM, FDtype = "AUC")
output4

}
\references{
Chao, A., Chiu C.-H. and Jost, L. (2010). Phylogenetic diversity measures based on Hill numbers. \emph{Philosophical Transactions of the Royal Society B.}, 365, 3599-3609. \cr\cr
Chao, A., Chiu, C.-H., Hsieh, T. C., Davis, T., Nipperess, D., and Faith, D. (2015). Rarefaction and extrapolation of phylogenetic diversity. \emph{Methods in Ecology and Evolution}, 6, 380-388.\cr\cr
Chao, A., Chiu C.-H. and Jost L. (2016). Phylogenetic diversity measures and their decomposition: a framework based on Hill numbers. pp. 141-172 in Pellens R. and Grandcolas P. (eds)
\emph{Biodiversity Conservation and Phylogenetic Systematics: Preserving our Evolutionary Heritage in an Extinction Crisis}, Springer. \cr\cr
Hsieh, T. C. and Chao, A. (2017). Rarefaction and extrapolation: making fair comparison of abundance-sensitive phylogenetic diversity among multiple assemblages. \emph{Systematic Biology}, 66, 100-111.
}
