% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{iNEXTBeta.link}
\alias{iNEXTBeta.link}
\title{Interpolation (rarefaction) and extrapolation of network Beta diversity
Function \code{iNEXTBeta.link} Interpolation and extrapolation of Beta diversity with order q}
\usage{
iNEXTBeta.link(
  data,
  diversity = "TD",
  level = seq(0.5, 1, 0.05),
  datatype = "abundance",
  q = c(0, 1, 2),
  nboot = 20,
  conf = 0.95,
  PDtype = "meanPD",
  row.tree = NULL,
  col.tree = NULL,
  row.distM = NULL,
  col.distM = NULL,
  FDtype = "AUC",
  FDtau = NULL,
  FDcut_number = 30
)
}
\arguments{
\item{data}{data can be input as a \code{lisst} of \code{data.frame}, each \code{data.frame} represents col.species-by-row.species abundance matrix; see Note 1 for an example.}

\item{diversity}{selection of diversity type: \code{'TD'} = 'Taxonomic diversity', \code{'PD'} = 'Phylogenetic diversity', and \code{'FD'} = 'Functional diversity'.}

\item{level}{a sequence specifying the particular sample coverages (between 0 and 1). Default is \code{seq(0.5, 1, 0.05)}.}

\item{q}{a numerical vector specifying the diversity orders. Default is \code{c(0,1,2)}.}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing
sampling uncertainty and constructing confidence intervals. Bootstrap replications are generally time consuming. Enter 0 to skip the bootstrap procedures. Default is \code{30}.}

\item{conf}{a positive number < 1 specifying the level of confidence interval. Default is \code{0.95}.}

\item{PDtype}{(required only when \code{diversity = "PD"}), select PD type: \code{PDtype = "PD"}(effective total branch length) or
\code{PDtype = "meanPD"}(effective number of equally divergent lineages).Default is \code{"meanPD"}.}

\item{row.tree}{(required only when \code{diversity = "PD"}), a phylogenetic tree of row assemblage in the pooled network row assemblage.}

\item{col.tree}{(required only when \code{diversity = "PD"}), a phylogenetic tree of column assemblage in the pooled network column assemblage.}

\item{row.distM}{(required only when \code{diversity = "FD"}), a species pairwise distance matrix for all species of row assemblage in the pooled network row assemblage.}

\item{col.distM}{(required only when \code{diversity = "FD"}), a species pairwise distance matrix for all species of column assemblage in the pooled network column assemblage.}

\item{FDtype}{(required only when \code{diversity = "FD"}), select FD type: \code{FDtype = "tau_values"} for FD under specified threshold values, or \code{FDtype = "AUC"} (area under the curve of tau-profile) for an overall FD which integrates all threshold values between zero and one. Default is \code{"AUC"}.}

\item{FDtau}{(required only when \code{diversity = "FD"} and \code{FDtype = "tau_values"}), a numerical vector between 0 and 1 specifying tau values (threshold levels). If \code{NULL} (default), then threshold is set to be the mean distance between any two individuals randomly selected from the pooled assemblage (i.e., quadratic entropy).}
}
\value{
A list of seven lists with three-diversity and four-dissimilarity.
}
\description{
Interpolation (rarefaction) and extrapolation of network Beta diversity
Function \code{iNEXTBeta.link} Interpolation and extrapolation of Beta diversity with order q
}
\examples{
## Taxonomic diversity
data(beetles)
output1 = iNEXTBeta.link(data = beetles, diversity = 'TD', level = seq(0.5, 0.9, 0.4), q = c(0, 1, 2))
output1

## Phylogenetic diversity
data(beetles)
data(beetles_col_tree)
output2 = iNEXTBeta.link(data = beetles, diversity = 'PD', level = seq(0.5, 0.9, 0.4), q = c(0, 1, 2), col.tree = beetles_col_tree)
output2


## Functional diversity under single threshold
data(beetles)
data(beetles_col_distM)
output3 = iNEXTBeta.link(data = beetles, diversity = 'FD', level = seq(0.5, 0.9, 0.4), q = c(0, 1, 2), col.distM = beetles_col_distM, FDtype = "tau_values")
output3


## Functional diversity with thresholds integrating from 0 to 1
data(beetles)
data(beetles_col_distM)
output4 = iNEXTBeta.link(data = beetles, diversity = 'FD', level = seq(0.5, 0.9, 0.4), q = c(0, 1, 2), col.distM = beetles_col_distM, FDtype = "AUC")
output4
}
\references{
Chao, A., Chazdon, R. L., Colwell, R. K. and Shen, T.-J.(2005). A new statistical approach for assessing similarity of species composition with incidence and abundance data. Ecology Letters 8, 148-159. (pdf file) Spanish translation in pp. 85-96 of Halffter, G. Soberon, J., Koleff, P. and Melic, A. (eds) 2005 Sobre Diversidad Biologica: el Sognificado de las Diversidades Alfa, Beta y Gamma. m3m-Monografias 3ercer Milenio, vol. 4, SEA, CONABIO, Grupo DIVERSITAS & CONACYT, Zaragoza. IV +242 pp.
Chiu, C.-H., Jost, L. and Chao*, A. (2014). Phylogenetic beta diversity, similarity, and differentiation measures based on Hill numbers. Ecological Monographs 84, 21-44.
}
