% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNEXTlink.R
\name{NetEvenness}
\alias{NetEvenness}
\title{Evenness Estimation of Evenness with order q}
\usage{
NetEvenness(
  x,
  q = seq(0, 2, 0.2),
  datatype = "abundance",
  method = "Estimated",
  nboot = 30,
  conf = 0.95,
  E.class = c(1:5),
  C = NULL
)
}
\arguments{
\item{outcome}{the outcome of the functions \code{ObsND} .\cr}
}
\value{
A list of estimated(empirical) evenness with order q.
Different lists represents different classes of Evenness.
Each list is combined with order.q and sites.
If "method" is estimated, then fist list will be named "C" which means the maximum standardized coverage between all double reference sample size.
}
\description{
\code{Evenness} computes Evenness Estimation of Evenness with order q.
}
