% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{ObsAsy.link}
\alias{ObsAsy.link}
\title{Asymptotic and Observed diversity q profile}
\usage{
ObsAsy.link(
  data,
  diversity = "TD",
  q = seq(0, 2, 0.2),
  nboot = 30,
  conf = 0.95,
  method = c("Asymptotic", "Observed"),
  row.tree = NULL,
  col.tree = NULL,
  PDtype = "meanPD",
  row.distM = NULL,
  col.distM = NULL,
  FDtype = "AUC",
  FDtau = NULL
)
}
\arguments{
\item{data}{a \code{list} of \code{data.frames}, each \code{data.frames} represents col.species-by-row.species abundance matrix.}

\item{diversity}{selection of diversity type: \code{'TD'} = Taxonomic network diversity, \code{'PD'} = Phylogenetic network diversity, and \code{'FD'} = Functional network diversity.}

\item{q}{a numerical vector specifying the diversity orders. Default is \code{seq(0, 2, by = 0.2)}.}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing sampling uncertainty and constructing confidence intervals. Enter 0 to skip the bootstrap procedures. Default is 50.}

\item{conf}{a positive number < 1 specifying the level of confidence interval. Default is 0.95.}

\item{method}{Select \code{'Asymptotic'} or \code{'Observed'}.}

\item{row.tree}{(required argument for \code{diversity = "PD"}), a phylogenetic tree in Newick format for row species in the row assemblage.}

\item{col.tree}{(required argument for \code{diversity = "PD"}), a phylogenetic tree in Newick format for column species in the column assemblage.}

\item{PDtype}{(argument only for \code{diversity = "PD"}), select PD type: \code{PDtype = "PD"} (effective total branch length) or \code{PDtype = "meanPD"} (effective number of equally divergent lineages). Default is \code{"meanPD"}, where \code{meanPD = PD/tree depth}.}

\item{row.distM}{(required argument for \code{diversity = "FD"}), a species pairwise distance matrix for row species in the row assemblage.}

\item{col.distM}{(required argument for \code{diversity = "FD"}), a species pairwise distance matrix for column species in the column assemblage.}

\item{FDtype}{(argument only for \code{diversity = "FD"}), select FD type: \code{FDtype = "tau_values"} for qiFD under specified threshold values, or \code{FDtype = "AUC"} (area under the curve of tau-profile) for an overall qiFD which integrates all threshold values between zero and one. Default is \code{"AUC"}.}

\item{FDtau}{(argument only for \code{diversity = "FD"} and \code{FDtype = "tau_values"}), a numerical vector between 0 and 1 specifying tau values (threshold levels). If \code{NULL} (default), then threshold is set to be the mean distance between any two individuals randomly selected from the pooled assemblage (i.e., quadratic entropy).}
}
\value{
a table of diversity table including the following arguments.
\item{Dataset}{the name of datasets.}
\item{Order.q}{the network diversity order of q.}
\item{qiTD, qiPD, qiFD}{the asymptotic/observed network network diversity estimates of order q.}
\item{s.e.}{standard error of the estimated network diversity.}
\item{qiTD.LCL, qiPD.LCL, qiFD.LCL and qiTD.UCL, qiPD.UCL, qiFD.UCL}{the bootstrap lower and upper confidence limits for the network diversity of order q at the specified level (with a default value of \code{0.95}).}
\item{Method}{\code{"Asymptotic"} means asymptotic network diversity and \code{"Observed"} means observed network diversity.}
\item{Reftime}{the reference times for qiPD.}
\item{Type}{\code{"PD"} (phylogenetic network diversity of effective total branch length),\code{"meanPD"} (effective number of equally divergent lineages) for qiPD.}
}
\description{
\code{ObsAsy.link}: The asymptotic (or observed) diversity of order q
}
\examples{
# Compute the observed and asymptotic taxonomic network diversity 
# for interaction data with order q between 0 and 2 
# (in increments of 0.2 by default)
data(beetles_plotA)
output_ObsAsy_qiTD = ObsAsy.link(data = beetles_plotA, diversity = 'TD', q = seq(0, 2, 0.2))
output_ObsAsy_qiTD

\donttest{
# Compute the observed and asymptotic phylogenetic network diversity
# for interaction data with order q between 0 and 2 
# (in increments of 0.2 by default)
data(beetles_plotA)
data(beetles_row_tree)
output_ObsAsy_qiPD = ObsAsy.link(data = beetles_plotA, diversity = 'PD', q = seq(0, 2, 0.2), 
                                 row.tree = beetles_row_tree, nboot = 10)
output_ObsAsy_qiPD


# Compute the observed and asymptotic functional network diversity
# for interaction data with order q between 0 and 2 and FDtype = "AUC"
# (in increments of 0.25 by default)
data(beetles_plotA)
data(beetles_row_distM)
output_ObsAsy_qiFD = ObsAsy.link(data = beetles_plotA, diversity = 'FD', q = seq(0, 2, 0.25), 
                                 row.distM = beetles_row_distM, FDtype = "AUC", nboot = 10)
output_ObsAsy_qiFD
}

}
