% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNEXTlink.R
\name{ObsND}
\alias{ObsND}
\title{Empirical diversity q profile}
\usage{
ObsND(
  data,
  q = seq(0, 2, 0.2),
  datatype = "abundance",
  nboot = 50,
  conf = 0.95
)
}
\arguments{
\item{outcome}{the outcome of the functions \code{ObsND} .\cr}
}
\value{
a table of Asymptotic network diversity q profile
}
\description{
\code{ObsND} The estimated and empirical diversity of order q
}
\examples{
\dontrun{
## Example for abundance-based data
data(Norfolk)
out1 <- ObsND(Norfolk, datatype = "abundance", nboot = 30)
ggObsND(out1)
}

}
