% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{Spec.link}
\alias{Spec.link}
\title{Standardized estimation (or observed) of specialization with order q}
\usage{
Spec.link(
  data,
  q = seq(0, 2, 0.2),
  method = "Estimated",
  nboot = 30,
  conf = 0.95,
  E.class = c(1:5),
  C = NULL
)
}
\arguments{
\item{data}{a \code{list} of \code{data.frames}, each \code{data.frames} represents col.species-by-row.species abundance matrix.}

\item{q}{a numerical vector specifying the diversity orders. Default is \code{seq(0, 2, 0.2)}.}

\item{method}{a binary calculation method with 'Estimated' or 'Observed'.}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing
sampling uncertainty and constructing confidence intervals. Bootstrap replications are generally time consuming. Enter 0 to skip the bootstrap procedures. Default is \code{30}.}

\item{conf}{a positive number < 1 specifying the level of confidence interval. Default is \code{0.95}.}

\item{E.class}{an integer vector between 1 to 5.}

\item{C}{a standardized coverage for calculating specialization index. It is used when \code{method = 'Estimated'}. If \code{NULL}, then this function computes the diversity estimates for the minimum sample coverage among all samples extrapolated to double reference sizes (\code{C = Cmax}).}
}
\value{
A list of several tables containing estimated (or observed) evenness with order q.\cr
        Each tables represents a class of specialization.\cr
        'Order.q' = the diversity order of q.\cr\cr
        'Specialization' = the specialization of order q.\cr\cr
        's.e.' = standard error of evenness.\cr\cr
        'Spec.LCL', 'Spec.UCL' = the bootstrap lower and upper confidence limits for the evenness of order q at the specified level (with a default value of 0.95).\cr\cr
        'Method' = "Estimated" or "Observed".\cr\cr
        'SC' = the target standardized coverage value. (only when method = 'Estimated') \cr\cr
        'Network' = the network name.\cr\cr
        'class' = specialization class.\cr\cr
}
\description{
Standardized estimation (or observed) of specialization with order q
}
\examples{
data(beetles)
output = Spec.link(beetles)
output
}
